+++ pym-go/utils.go
@@ -1,0 +1,103 @@
+package main
+
+import (
+	"crypto/rand"
+	"crypto/sha1"
+	"database/sql"
+	"encoding/hex"
+	"fmt"
+	"io"
+	"log"
+	"mime/multipart"
+	"time"
+)
+
+type handler struct {
+	db          *sql.DB
+	maxFileSize int64
+	maxTries    int64
+}
+
+type Body struct {
+	Group    string
+	Language string
+	Value    string
+}
+
+type Form struct {
+	File *multipart.FileHeader `form:"files"`
+}
+
+// hashFile returns the SHA1 hash of the given content
+func hashFile(file multipart.File) (string, error) {
+	hashSum := sha1.New()
+	if _, err := io.Copy(hashSum, file); err != nil {
+		log.Fatal(err)
+		return "", err
+	}
+	return fmt.Sprintf("%x", hashSum.Sum(nil)), nil
+}
+
+func (self *handler) updateExpire(shortId string) error {
+	newExpire := time.Now().AddDate(0, 0, 15).Format(time.RFC3339)
+	_, err := self.db.Exec(`UPDATE pym SET expire=$1 WHERE shortId=$2`, newExpire, shortId)
+	return err
+}
+
+func (self *handler) fetchPost(shortId string) (string, string, string, error) {
+	var hash string
+	var group string
+	var language string
+
+	row := self.db.QueryRow(`SELECT "group", language, hash FROM pym WHERE shortId=$1`, shortId)
+	err := row.Scan(&group, &language, &hash)
+
+	return hash, group, language, err
+}
+func hashText(buffer string) string {
+	return fmt.Sprintf("%x", sha1.Sum([]byte(buffer)))
+}
+
+func (self *handler) generateShortID() (string, error) {
+	var count int
+	var shortId string
+	b := make([]byte, 2)
+	for i := int64(0); i <= self.maxTries; i += 1 {
+		_, err := rand.Read(b)
+		if err != nil {
+			log.Println(err)
+			continue
+		}
+
+		shortId = hex.EncodeToString(b)[:4]
+		row := self.db.QueryRow(`SELECT COUNT(shortId) FROM pym WHERE shortId=$1`, shortId)
+		err = row.Scan(&count)
+		if err != nil {
+			log.Println(err)
+		}
+		if count == 0 {
+			break
+		}
+	}
+
+	// If we are out of the loop and the count is not 0, then we could not make new entry
+	if count != 0 {
+		return "", fmt.Errorf("Could not make new entry in database.")
+	}
+	return shortId, nil
+}
+
+func (self *handler) checkHash(hash string) (string, error) {
+	var shortId string
+	row := self.db.QueryRow(`SELECT shortId FROM pym WHERE hash=$1`, hash)
+	err := row.Scan(&shortId)
+
+	switch err {
+	case nil:
+		return shortId, nil
+	case sql.ErrNoRows:
+		return "", nil
+	default:
+		return "", err
+	}
+}
